requirements.txt
pyautogui==0.9.53
Pillow~=9.2.0